% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putior.R
\name{as_putior_workflow}
\alias{as_putior_workflow}
\title{Add putior_workflow class to a data frame}
\usage{
as_putior_workflow(df)
}
\arguments{
\item{df}{Data frame to class}
}
\value{
Data frame with "putior_workflow" class prepended
}
\description{
Add putior_workflow class to a data frame
}
\keyword{internal}
