% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putior.R
\name{convert_results_to_df}
\alias{convert_results_to_df}
\title{Convert results list to data frame}
\usage{
convert_results_to_df(results, include_line_numbers)
}
\arguments{
\item{results}{List of annotation results}

\item{include_line_numbers}{Whether line numbers are included}
}
\value{
Data frame
}
\description{
Convert results list to data frame
}
\keyword{internal}
