% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{create_artifact_nodes}
\alias{create_artifact_nodes}
\title{Create artifact nodes for data files}
\usage{
create_artifact_nodes(workflow)
}
\arguments{
\item{workflow}{Workflow data frame}
}
\value{
Data frame with artifact node definitions
}
\description{
Create artifact nodes for data files
}
\keyword{internal}
