% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{detect_files_from_patterns}
\alias{detect_files_from_patterns}
\title{Detect files from pattern list}
\usage{
detect_files_from_patterns(content, lines, pattern_list)
}
\arguments{
\item{content}{Full file content as single string}

\item{lines}{File content as line vector}

\item{pattern_list}{List of detection patterns}
}
\value{
List with files and lines vectors
}
\description{
Detect files from pattern list
}
\keyword{internal}
