% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{detect_workflow_elements}
\alias{detect_workflow_elements}
\title{Detect workflow elements in a single file}
\usage{
detect_workflow_elements(
  file,
  detect_inputs = TRUE,
  detect_outputs = TRUE,
  detect_dependencies = TRUE,
  include_line_numbers = FALSE
)
}
\arguments{
\item{file}{Path to source file}

\item{detect_inputs}{Whether to detect inputs}

\item{detect_outputs}{Whether to detect outputs}

\item{detect_dependencies}{Whether to detect dependencies}

\item{include_line_numbers}{Whether to include line numbers}
}
\value{
Named list with workflow elements
}
\description{
Detect workflow elements in a single file
}
\keyword{internal}
