% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{extract_file_path_from_line}
\alias{extract_file_path_from_line}
\title{Extract file path from a line containing a matched function call}
\usage{
extract_file_path_from_line(line, pattern_def)
}
\arguments{
\item{line}{The line containing the function call}

\item{pattern_def}{The pattern definition}
}
\value{
Character vector of extracted file paths
}
\description{
Extract file path from a line containing a matched function call
}
\keyword{internal}
