% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{generate_annotation_for_file}
\alias{generate_annotation_for_file}
\title{Generate annotation text for a single file}
\usage{
generate_annotation_for_file(file, style = "multiline")
}
\arguments{
\item{file}{Path to source file}

\item{style}{Annotation style ("single" or "multiline")}
}
\value{
Annotation string
}
\description{
Generate annotation text for a single file
}
\keyword{internal}
