% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{generate_connections}
\alias{generate_connections}
\title{Generate connections between nodes}
\usage{
generate_connections(workflow, show_files = FALSE)
}
\arguments{
\item{workflow}{Workflow data frame}

\item{show_files}{Whether to show file-based connections}
}
\value{
Character vector of connection definitions
}
\description{
Generate connections between nodes
}
\keyword{internal}
