% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{generate_node_definitions}
\alias{generate_node_definitions}
\title{Generate node definitions for mermaid diagram}
\usage{
generate_node_definitions(
  workflow,
  node_labels = "label",
  show_workflow_boundaries = TRUE,
  show_source_info = FALSE
)
}
\arguments{
\item{workflow}{Workflow data frame}

\item{node_labels}{What to show in node labels}

\item{show_workflow_boundaries}{Whether to apply special styling to start/end nodes}

\item{show_source_info}{Whether to append source file info to node labels}
}
\value{
Character vector of node definitions
}
\description{
Generate node definitions for mermaid diagram
}
\keyword{internal}
