% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_c_patterns}
\alias{get_c_patterns}
\title{Get C Detection Patterns}
\usage{
get_c_patterns()
}
\value{
List of C detection patterns
}
\description{
Get C Detection Patterns
}
\keyword{internal}
