% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_registry.R
\name{get_comment_group}
\alias{get_comment_group}
\title{Get Comment Syntax Group for Extension}
\usage{
get_comment_group(ext)
}
\arguments{
\item{ext}{Character string of the file extension (without dot)}
}
\value{
Character string of the group name ("hash", "dash", "slash", "percent"),
or "hash" if not found (default)
}
\description{
Returns the name of the comment syntax group for a given file extension.
}
\keyword{internal}
