% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_cpp_patterns}
\alias{get_cpp_patterns}
\title{Get C++ Detection Patterns}
\usage{
get_cpp_patterns()
}
\value{
List of C++ detection patterns
}
\description{
Get C++ Detection Patterns
}
\keyword{internal}
