% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_dockerfile_patterns}
\alias{get_dockerfile_patterns}
\title{Get Dockerfile Detection Patterns}
\usage{
get_dockerfile_patterns()
}
\value{
List of Dockerfile detection patterns
}
\description{
Get Dockerfile Detection Patterns
}
\keyword{internal}
