% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_go_patterns}
\alias{get_go_patterns}
\title{Get Go Detection Patterns}
\usage{
get_go_patterns()
}
\value{
List of Go detection patterns
}
\description{
Get Go Detection Patterns
}
\keyword{internal}
