% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_java_patterns}
\alias{get_java_patterns}
\title{Get Java Detection Patterns}
\usage{
get_java_patterns()
}
\value{
List of Java detection patterns
}
\description{
Get Java Detection Patterns
}
\keyword{internal}
