% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_javascript_patterns}
\alias{get_javascript_patterns}
\title{Get JavaScript Detection Patterns}
\usage{
get_javascript_patterns()
}
\value{
List of JavaScript detection patterns
}
\description{
Get JavaScript Detection Patterns
}
\keyword{internal}
