% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_julia_patterns}
\alias{get_julia_patterns}
\title{Get Julia Detection Patterns}
\usage{
get_julia_patterns()
}
\value{
List of Julia detection patterns
}
\description{
Get Julia Detection Patterns
}
\keyword{internal}
