% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_lua_patterns}
\alias{get_lua_patterns}
\title{Get Lua Detection Patterns}
\usage{
get_lua_patterns()
}
\value{
List of Lua detection patterns
}
\description{
Get Lua Detection Patterns
}
\keyword{internal}
