% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_makefile_patterns}
\alias{get_makefile_patterns}
\title{Get Makefile Detection Patterns}
\usage{
get_makefile_patterns()
}
\value{
List of Makefile detection patterns
}
\description{
Get Makefile Detection Patterns
}
\keyword{internal}
