% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_matlab_patterns}
\alias{get_matlab_patterns}
\title{Get MATLAB Detection Patterns}
\usage{
get_matlab_patterns()
}
\value{
List of MATLAB detection patterns
}
\description{
Get MATLAB Detection Patterns
}
\keyword{internal}
