% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_python_patterns}
\alias{get_python_patterns}
\title{Get Python Detection Patterns}
\usage{
get_python_patterns()
}
\value{
List of Python detection patterns
}
\description{
Get Python Detection Patterns
}
\keyword{internal}
