% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_r_patterns}
\alias{get_r_patterns}
\title{Get R Detection Patterns}
\usage{
get_r_patterns()
}
\value{
List of R detection patterns
}
\description{
Get R Detection Patterns
}
\keyword{internal}
