% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_ruby_patterns}
\alias{get_ruby_patterns}
\title{Get Ruby Detection Patterns}
\usage{
get_ruby_patterns()
}
\value{
List of Ruby detection patterns
}
\description{
Get Ruby Detection Patterns
}
\keyword{internal}
