% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_rust_patterns}
\alias{get_rust_patterns}
\title{Get Rust Detection Patterns}
\usage{
get_rust_patterns()
}
\value{
List of Rust detection patterns
}
\description{
Get Rust Detection Patterns
}
\keyword{internal}
