% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_shell_patterns}
\alias{get_shell_patterns}
\title{Get Shell Detection Patterns}
\usage{
get_shell_patterns()
}
\value{
List of shell detection patterns
}
\description{
Get Shell Detection Patterns
}
\keyword{internal}
