% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_sql_patterns}
\alias{get_sql_patterns}
\title{Get SQL Detection Patterns}
\usage{
get_sql_patterns()
}
\value{
List of SQL detection patterns
}
\description{
Get SQL Detection Patterns
}
\keyword{internal}
