% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_registry.R
\name{get_supported_extensions}
\alias{get_supported_extensions}
\title{Get All Supported File Extensions}
\usage{
get_supported_extensions()
}
\value{
Character vector of supported file extensions (without dots)
}
\description{
Returns a character vector of all file extensions supported by putior's
annotation parsing system.
}
\examples{
get_supported_extensions()
}
\seealso{
Other language-support: 
\code{\link{ext_to_language}()},
\code{\link{get_comment_prefix}()},
\code{\link{get_detection_patterns}()},
\code{\link{list_supported_languages}()}
}
\concept{language-support}
