% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{get_theme_colors}
\alias{get_theme_colors}
\title{Get color schemes for different themes (FIXED VERSION)}
\usage{
get_theme_colors(theme)
}
\arguments{
\item{theme}{Theme name}
}
\value{
Named list of color definitions for each node type
}
\description{
Get color schemes for different themes (FIXED VERSION)
}
\keyword{internal}
