% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_patterns.R
\name{get_wgsl_patterns}
\alias{get_wgsl_patterns}
\title{Get WGSL Detection Patterns}
\usage{
get_wgsl_patterns()
}
\value{
List of WGSL detection patterns
}
\description{
Get WGSL Detection Patterns
}
\keyword{internal}
