% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{infer_node_type}
\alias{infer_node_type}
\title{Infer node type from inputs and outputs}
\usage{
infer_node_type(inputs, outputs)
}
\arguments{
\item{inputs}{Detected input files}

\item{outputs}{Detected output files}
}
\value{
Character string of node type
}
\description{
Infer node type from inputs and outputs
}
\keyword{internal}
