% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{insert_annotation_into_file}
\alias{insert_annotation_into_file}
\title{Insert annotation at top of file}
\usage{
insert_annotation_into_file(file, annotation)
}
\arguments{
\item{file}{Path to source file}

\item{annotation}{Annotation text to insert}
}
\description{
Insert annotation at top of file
}
\keyword{internal}
