% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{is_likely_file_path}
\alias{is_likely_file_path}
\title{Check if a string looks like a file path}
\usage{
is_likely_file_path(str)
}
\arguments{
\item{str}{String to check}
}
\value{
Logical
}
\description{
Check if a string looks like a file path
}
\keyword{internal}
