% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{merge_annotations}
\alias{merge_annotations}
\title{Merge manual and auto-detected annotations}
\usage{
merge_annotations(manual, auto, strategy)
}
\arguments{
\item{manual}{Data frame of manual annotations}

\item{auto}{Data frame of auto-detected annotations}

\item{strategy}{Merge strategy}
}
\value{
Merged data frame
}
\description{
Merge manual and auto-detected annotations
}
\keyword{internal}
