% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putior.R
\name{parse_comma_separated_pairs}
\alias{parse_comma_separated_pairs}
\title{Parse comma-separated pairs while respecting quotes}
\usage{
parse_comma_separated_pairs(text)
}
\arguments{
\item{text}{Text to parse}
}
\value{
Character vector of pairs
}
\description{
Parse comma-separated pairs while respecting quotes
}
\keyword{internal}
