% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putior.R
\name{print.putior_workflow}
\alias{print.putior_workflow}
\title{Print a putior workflow}
\usage{
\method{print}{putior_workflow}(x, ...)
}
\arguments{
\item{x}{A putior_workflow object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Displays a concise summary of a putior workflow data frame.
}
\seealso{
Other core-workflow: 
\code{\link{print.putior_theme}()},
\code{\link{put}()},
\code{\link{put_diagram}()},
\code{\link{put_theme}()},
\code{\link{summary.putior_workflow}()}
}
\concept{core-workflow}
