% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putior.R
\name{process_single_file}
\alias{process_single_file}
\title{Process a single file for PUT annotations}
\usage{
process_single_file(file, include_line_numbers, validate)
}
\arguments{
\item{file}{Path to file}

\item{include_line_numbers}{Whether to include line numbers}

\item{validate}{Whether to validate annotations}
}
\value{
List of annotation results or error message
}
\description{
Process a single file for PUT annotations
}
\keyword{internal}
