% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{putior-package}
\alias{putior-package}
\title{putior Package Initialization}
\description{
Package load and attach hooks for putior. Sets up default options
and initializes the logging namespace when the logger package is available.
}
\keyword{internal}
