% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{sanitize_node_id}
\alias{sanitize_node_id}
\title{Sanitize node ID for mermaid compatibility}
\usage{
sanitize_node_id(name)
}
\arguments{
\item{name}{Node name}
}
\value{
Sanitized node ID
}
\description{
Sanitize node ID for mermaid compatibility
}
\keyword{internal}
