% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{split_file_list}
\alias{split_file_list}
\title{Split comma-separated file list}
\usage{
split_file_list(file_string)
}
\arguments{
\item{file_string}{Comma-separated file names}
}
\value{
Character vector of individual file names
}
\description{
Split comma-separated file list
}
\keyword{internal}
