% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_auto.R
\name{ext_to_language}
\alias{ext_to_language}
\title{Convert file extension to language identifier}
\usage{
ext_to_language(ext)
}
\arguments{
\item{ext}{File extension}
}
\value{
Language identifier or NULL
}
\description{
Convert file extension to language identifier
}
\keyword{internal}
