% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{generate_node_definitions}
\alias{generate_node_definitions}
\title{Generate node definitions for mermaid diagram}
\usage{
generate_node_definitions(workflow, node_labels)
}
\arguments{
\item{workflow}{Workflow data frame}

\item{node_labels}{What to show in node labels}
}
\value{
Character vector of node definitions
}
\description{
Generate node definitions for mermaid diagram
}
\keyword{internal}
