% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{generate_node_styling}
\alias{generate_node_styling}
\title{Generate node styling based on node types}
\usage{
generate_node_styling(workflow)
}
\arguments{
\item{workflow}{Workflow data frame}
}
\value{
Character vector of styling definitions
}
\description{
Generate node styling based on node types
}
\keyword{internal}
