% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diagram.R
\name{handle_output}
\alias{handle_output}
\title{Handle output of mermaid diagram}
\usage{
handle_output(mermaid_code, output, file, title)
}
\arguments{
\item{mermaid_code}{Generated mermaid code}

\item{output}{Output method}

\item{file}{File path for file output}

\item{title}{Diagram title}
}
\description{
Handle output of mermaid diagram
}
\keyword{internal}
