% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_theme.R
\name{print.putior_theme}
\alias{print.putior_theme}
\title{Print a putior_theme Object}
\usage{
\method{print}{putior_theme}(x, ...)
}
\arguments{
\item{x}{A \code{putior_theme} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object, invisibly.
}
\description{
Print a putior_theme Object
}
